Fixed bugs and refactored code (#15)
* Remove logging douyin danmu data

* [douyin] add a way to generate msToken

* [danmu] fix danmu end of line not being appended

* [refactor] remove some unused dependencies

* [refactor] reduce douyin ttwid and msToken logs

* [danmu] support multiple network frames in a single request

* Refactor logging configuration to be programmatically set

The previous logging configuration file (logback.xml) has been removed, and the logging configuration functionality has instead been implemented directly within the Application.kt class. This approach has been chosen to allow for easier manipulation and visibility of logging settings directly within the main codebase. It also includes environmental overrides and console log and file logging setup.

* Set up process directory for commands actions

When creating a process of a command action, the output folder specified in the application configuration will now be used as the process's directory if it is not empty.

* Refactor kotlinx.coroutines dependencies and update libraries versions

The build.gradle.kts file was updated to include kotlinx.coroutines dependencies. Correspondingly, the versions for these new additions were defined in the libs.versions.toml file.

* Remove 'kotlin-xml-builder' dependency and refactor Danmu XML creation