[dev] refactor and bugfixes (#16)
* [danmu] ensure that each danmu write channel has only one onClose handler

* [refactor] refactor code

* [app] monitor changes of toml configuration.

* [fix] when a ping timeout happens when receiving danmus, it cancels its parent job.

Added a backoff strategic logic which allows danmu's retry to be capped at 10 times with an initial delay of 10 seconds. The maximum delay will be 60 seconds with a 1.5 factor.

* Create docker-image.yml

* [danmu] hide some logs

* [download] refactor download code

* [download] add maxPartDuration functionality to download segmenting

Introduced a new configuration parameter, maxPartDuration, to control the duration of download parts in addition to the existing size-based splitting. If a maxPartDuration value is set, it will prioritize the time split over the size split.